fix warning

+++ cocos/ui/UIListView.cpp
@@ -806,7 +806,7 @@ ssize_t ListView::getCurSelectedIndex() const
     return _curSelectedIndex;
 }
 
-void ListView::setCurSelectedIndex(int itemIndex)
+void ListView::setCurSelectedIndex(size_t itemIndex)
 {
     Widget* item = getItem(itemIndex);
     if (item == nullptr)
